Implement Basic Thread API (#11)
* create tests for thread commands

* implement spawning threads

* bit of refactoring

* implement thread exit

* add system calls to get current pid and tid

* implement WatchThread and make sure weird entry points avoid kernel panic

+++ kernel/src/process/mod.rs
@@ -10,19 +10,17 @@ use crate::{
     },
     tasks::locks::Mutex,
 };
-use alloc::{sync::Arc, vec::Vec};
+use alloc::sync::Arc;
 
 use bitfield::Bit;
 use core::{
     num::NonZeroU32,
     sync::atomic::{AtomicU16, AtomicU32},
-    task::Waker,
 };
-use futures::Future;
 use kapi::{commands::Command, completions::Completion};
 use smallvec::SmallVec;
 use snafu::{OptionExt, ResultExt};
-use spin::{Mutex as SpinMutex, Once};
+use spin::Once;
 
 pub mod thread;
 pub use thread::{scheduler::scheduler, threads, Thread, ThreadId};
@@ -41,11 +39,12 @@ pub use kapi::ProcessId;
 /// A user-space process.
 ///
 /// A process is a collection of threads that share the same address space and system resources.
+/// For a process to end, *all* of its associated threads must first exit.
 pub struct Process {
     /// The ID of this process.
     pub id: ProcessId,
     /// The threads running in this process.
-    threads: Mutex<SmallVec<[Arc<Thread>; 2]>>,
+    pub threads: Mutex<SmallVec<[Arc<Thread>; 2]>>,
     /// Page tables for this process.
     page_tables: PageTable,
     /// Allocator for the virtual address space of the process.
@@ -58,11 +57,9 @@ pub struct Process {
     #[allow(clippy::type_complexity)]
     send_queues: ConcurrentLinkedList<(Arc<OwnedQueue<Command>>, Arc<OwnedQueue<Completion>>)>,
     /// The receive (kernel to user space) queues associated with this process.
+    // TODO: these queues don't need to be owned, because the process implicitly owns
+    // all the mapped memory in its page tables by default.
     recv_queues: ConcurrentLinkedList<Arc<OwnedQueue<Completion>>>,
-    /// The exit code this process exited with (if it has exited).
-    exit_code: Once<Option<u32>>,
-    /// Future wakers for futures waiting on exit code.
-    exit_waker: SpinMutex<Vec<Waker>>,
 }
 
 crate::assert_sync!(Process);
@@ -78,18 +75,21 @@ impl Process {
         // downgrade the Arc so that pending tasks don't keep the process alive unnecessarily
         let this = &Arc::downgrade(self);
         for (send_qu, assoc_recv_qu) in self.send_queues.iter() {
+            let assoc_recv_qu = Arc::downgrade(assoc_recv_qu);
             while let Some(cmd) = send_qu.poll() {
                 let this = this.clone();
                 let assoc_recv_qu = assoc_recv_qu.clone();
                 log::trace!("[pid {}, SQ {}]: {cmd:?}", self.id, send_qu.id);
                 crate::tasks::spawn(async move {
                     if let Some(proc) = this.upgrade() {
-                        let cmpl = interface::dispatch(&proc, cmd).await;
-                        log::trace!("[pid {}, RQ {}]: {cmpl:?}", proc.id, assoc_recv_qu.id);
-                        match assoc_recv_qu.post(cmpl) {
-                            Ok(()) => {}
-                            Err(_) => {
-                                todo!("kill process on queue overflow")
+                        let cmpl = proc.dispatch_command(cmd, assoc_recv_qu.clone()).await;
+                        if let Some(arq) = assoc_recv_qu.upgrade() {
+                            log::trace!("[pid {}, RQ {}]: {cmpl:?}", proc.id, arq.id);
+                            match arq.post(cmpl) {
+                                Ok(()) => {}
+                                Err(_) => {
+                                    todo!("kill process on queue overflow")
+                                }
                             }
                         }
                     }
@@ -107,15 +107,17 @@ impl Process {
             self.id,
             thread.id
         );
-        // TODO: the thread needs to go into some kind of dead/error state and/or the process
-        // needs to be notified that one of its threads just died. If this is the last thread
-        // in the process than the process itself is now dead
-    }
 
-    /// Create a future that will resolve with the value of the exit code when this process exits.
-    /// If the process is killed or aborted, then [None] will be returned.
-    pub fn exit_code(self: &Arc<Process>) -> impl Future<Output = Option<u32>> {
-        ProcessExitFuture { proc: self.clone() }
+        thread.exit(kapi::completions::ThreadExit::PageFault);
+    }
+}
+
+impl Drop for Process {
+    fn drop(&mut self) {
+        for mapped_region in self.page_tables.iter() {
+            physical_memory_allocator()
+                .free_pages(mapped_region.base_phys_addr, mapped_region.page_count);
+        }
     }
 }
 
@@ -132,63 +134,33 @@ pub fn process_for_id(id: ProcessId) -> Option<Arc<Process>> {
     processes().iter().find(|p| p.id == id).cloned()
 }
 
-struct ProcessExitFuture {
-    proc: Arc<Process>,
-}
-
-impl Future for ProcessExitFuture {
-    type Output = Option<u32>;
-
-    fn poll(
-        self: core::pin::Pin<&mut Self>,
-        cx: &mut core::task::Context<'_>,
-    ) -> core::task::Poll<Self::Output> {
-        match self.proc.exit_code.poll() {
-            Some(c) => core::task::Poll::Ready(*c),
-            None => {
-                self.proc.exit_waker.lock().push(cx.waker().clone());
-                core::task::Poll::Pending
-            }
-        }
-    }
-}
-
-/// Cause a process to exit, freeing its resources.
-///
-/// Because threads have strong references to their parent process, process resources will not be
-/// freed unless this function is called.
-// TODO: this should be somewhere else, perhaps with `spawn`.
-fn exit_process(proc: Arc<Process>, exit_code: Option<u32>) {
-    log::trace!("process {} exited with code {exit_code:?}", proc.id);
-
-    processes().remove(|p| p.id == proc.id);
-
-    // delete and unschedule all threads
-    // removing the current thread will automatically schedule a new thread to run
-    for thread in proc.threads.lock_blocking().drain(..) {
-        scheduler().remove_thread(&thread);
-        threads().remove(|t| t.id == thread.id);
-        // drop thread
-    }
-
-    proc.exit_code.call_once(|| exit_code);
-    for w in proc.exit_waker.lock().drain(..) {
-        w.wake();
-    }
-
-    for mapped_region in proc.page_tables.iter() {
-        physical_memory_allocator()
-            .free_pages(mapped_region.base_phys_addr, mapped_region.page_count);
-    }
-}
-
 /// Register system calls related to processes and thread scheduling.
 pub fn register_system_call_handlers() {
     use kapi::system_calls::SystemCallNumber;
     let h = crate::exception::system_call_handlers();
-    h.insert_blocking(SystemCallNumber::Exit as u16, |_id, proc, _thread, regs| {
-        exit_process(proc, Some(regs.x[0] as u32));
+    h.insert_blocking(SystemCallNumber::Exit as u16, |_id, _proc, thread, regs| {
+        thread.exit(kapi::completions::ThreadExit::Normal(regs.x[0] as u16));
     });
+    h.insert_blocking(
+        SystemCallNumber::GetCurrentProcessId as u16,
+        |_id, proc, _thread, regs| {
+            // TODO: validate pointer
+            let p = regs.x[0] as *mut NonZeroU32;
+            unsafe {
+                p.write(proc.id);
+            }
+        },
+    );
+    h.insert_blocking(
+        SystemCallNumber::GetCurrentThreadId as u16,
+        |_id, _proc, thread, regs| {
+            // TODO: validate pointer
+            let p = regs.x[0] as *mut u32;
+            unsafe {
+                p.write(thread.id);
+            }
+        },
+    );
     h.insert_blocking(
         SystemCallNumber::Yield as u16,
         |_id, _proc, _thread, _regs| {
